Separate runDevTools so that it can be used safely in both contexts
